Code formatting, test tightening, inline comment cleanup.
The eclipse setting to convert tabs to spaces had been created but not
enabled for this project. It is now enabled.

One unit test had assertion code that predated the wrapper classes for
MapWritable and ArrayWritable, which now can be compared correctly at
their respective object levels.

Fixed misplaced comments and edited them for clarity.